+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ### Oneview Features supported
 - Enclosure
 - Enclosure Group
+- Hypervisor Cluster Profile
+- Hypervisor Manager
+- Server Certificate
 - Uplink Sets
 
 #### Bug fixes & Enhancements: